Refactor: Migrate from ESLint to oxlint
- Replace ESLint with oxlint (Rust-based linter)
- Enforce default exports only (import/no-named-export)
- Enforce no barrel files (oxc/no-barrel-file)
- Split multi-type files into single-export files
- Remove 210 ESLint-related packages
- Update lint-staged to use oxlint

Benefits:
- ~10x faster linting (70ms for 72 files)
- Better module loading performance with tsx
- Simpler, more focused lint rules
