--- .eslintrc.cjs
@@ -1,27 +1,0 @@
-module.exports = {
-  root: true,
-  parser: "@typescript-eslint/parser",
-  parserOptions: {
-    project: "./tsconfig.json",
-  },
-  plugins: ["@typescript-eslint", "import"],
-  extends: [
-    "airbnb-typescript/base",
-    "plugin:@typescript-eslint/recommended",
-    "prettier",
-  ],
-  rules: {
-    // Allow console for CLI tool
-    "no-console": "off",
-    // Allow underscore-prefixed variables to be unused (common pattern for ignored params)
-    "@typescript-eslint/no-unused-vars": [
-      "error",
-      {
-        argsIgnorePattern: "^_",
-        varsIgnorePattern: "^_",
-        caughtErrorsIgnorePattern: "^_",
-      },
-    ],
-  },
-  ignorePatterns: ["dist/", "node_modules/", "src/parser/", "*.js"],
-};
