+++ .oxlintrc.json
@@ -1,0 +1,27 @@
+{
+  "$schema": "./node_modules/oxlint/configuration_schema.json",
+  "plugins": ["import", "oxc", "typescript"],
+  "categories": {
+    "correctness": "error",
+    "suspicious": "warn"
+  },
+  "rules": {
+    "no-console": "off",
+    "@typescript-eslint/no-unused-vars": [
+      "error",
+      {
+        "argsIgnorePattern": "^_",
+        "varsIgnorePattern": "^_",
+        "caughtErrorsIgnorePattern": "^_"
+      }
+    ],
+    "import/no-named-export": "error",
+    "import/group-exports": "off",
+    "@typescript-eslint/consistent-type-imports": "off",
+    "@typescript-eslint/no-inferrable-types": "off",
+    "@typescript-eslint/prefer-for-of": "off",
+    "@typescript-eslint/no-extraneous-class": "off",
+    "oxc/no-barrel-file": "error"
+  },
+  "ignorePatterns": ["dist/**", "node_modules/**", "src/parser/**", "**/*.js"]
+}
