+++ src/codegen/types/BITMAP_SIZE.ts
@@ -1,0 +1,11 @@
+/**
+ * ADR-034: Bitmap type sizes (total bits)
+ */
+const BITMAP_SIZE: Record<string, number> = {
+  bitmap8: 8,
+  bitmap16: 16,
+  bitmap24: 24,
+  bitmap32: 32,
+};
+
+export default BITMAP_SIZE;
