+++ src/codegen/types/FLOAT_TYPES.ts
@@ -1,0 +1,6 @@
+/**
+ * ADR-024: Floating point types
+ */
+const FLOAT_TYPES = ["f32", "f64"] as const;
+
+export default FLOAT_TYPES;
