+++ src/codegen/types/IAssignmentContext.ts
@@ -1,0 +1,12 @@
+import TOverflowBehavior from "./TOverflowBehavior";
+
+/**
+ * Assignment context for overflow behavior tracking
+ */
+interface IAssignmentContext {
+  targetName: string | null;
+  targetType: string | null;
+  overflowBehavior: TOverflowBehavior;
+}
+
+export default IAssignmentContext;
