+++ src/codegen/types/SIGNED_TYPES.ts
@@ -1,0 +1,6 @@
+/**
+ * ADR-024: Signed integer types
+ */
+const SIGNED_TYPES = ["i8", "i16", "i32", "i64"] as const;
+
+export default SIGNED_TYPES;
