+++ src/codegen/types/TYPE_RANGES.ts
@@ -1,0 +1,16 @@
+/**
+ * ADR-024: Type ranges for literal validation
+ * Maps type name to [min, max] inclusive range
+ */
+const TYPE_RANGES: Record<string, [bigint, bigint]> = {
+  u8: [0n, 255n],
+  u16: [0n, 65535n],
+  u32: [0n, 4294967295n],
+  u64: [0n, 18446744073709551615n],
+  i8: [-128n, 127n],
+  i16: [-32768n, 32767n],
+  i32: [-2147483648n, 2147483647n],
+  i64: [-9223372036854775808n, 9223372036854775807n],
+};
+
+export default TYPE_RANGES;
