+++ src/codegen/types/TYPE_WIDTH.ts
@@ -1,0 +1,18 @@
+/**
+ * Maps primitive types to their bit widths
+ */
+const TYPE_WIDTH: Record<string, number> = {
+  u8: 8,
+  i8: 8,
+  u16: 16,
+  i16: 16,
+  u32: 32,
+  i32: 32,
+  u64: 64,
+  i64: 64,
+  f32: 32,
+  f64: 64,
+  bool: 1,
+};
+
+export default TYPE_WIDTH;
