+++ src/codegen/types/UNSIGNED_TYPES.ts
@@ -1,0 +1,6 @@
+/**
+ * ADR-024: Unsigned integer types
+ */
+const UNSIGNED_TYPES = ["u8", "u16", "u32", "u64"] as const;
+
+export default UNSIGNED_TYPES;
