+++ src/lib/transpiler.ts
@@ -8,39 +8,13 @@ import { CNextLexer } from "../parser/grammar/CNextLexer";
 import { CNextParser } from "../parser/grammar/CNextParser";
 import CodeGenerator from "../codegen/CodeGenerator";
 import CommentExtractor from "../codegen/CommentExtractor";
-import CNextSymbolCollector from "../symbols/CNextSymbolCollector";
-import ESymbolKind from "../types/ESymbolKind";
 import InitializationAnalyzer from "../analysis/InitializationAnalyzer";
 import FunctionCallAnalyzer from "../analysis/FunctionCallAnalyzer";
 import NullCheckAnalyzer from "../analysis/NullCheckAnalyzer";
 import DivisionByZeroAnalyzer from "../analysis/DivisionByZeroAnalyzer";
-import {
-  ITranspileResult,
-  ITranspileError,
-  ISymbolInfo,
-  IParseWithSymbolsResult,
-  TSymbolKind,
-} from "./types/ITranspileResult";
-
-export {
-  ITranspileResult,
-  ITranspileError,
-  ISymbolInfo,
-  IParseWithSymbolsResult,
-  TSymbolKind,
-};
-
-/**
- * Options for transpilation
- */
-export interface ITranspileOptions {
-  /** Parse only, don't generate code */
-  parseOnly?: boolean;
-  /** ADR-044: When true, generate panic-on-overflow helpers instead of clamp helpers */
-  debugMode?: boolean;
-  /** ADR-049: Target platform for atomic code generation (e.g., "teensy41", "cortex-m0") */
-  target?: string;
-}
+import ITranspileResult from "./types/ITranspileResult";
+import ITranspileError from "./types/ITranspileError";
+import ITranspileOptions from "./types/ITranspileOptions";
 
 /**
  * Transpile C-Next source code to C
@@ -51,7 +25,7 @@ export interface ITranspileOptions {
  *
  * @example
  * ```typescript
- * import { transpile } from './lib/transpiler';
+ * import transpile from './lib/transpiler';
  *
  * const result = transpile('u32 x <- 5;');
  * if (result.success) {
@@ -61,7 +35,7 @@ export interface ITranspileOptions {
  * }
  * ```
  */
-export function transpile(
+function transpile(
   source: string,
   options: ITranspileOptions = {},
 ): ITranspileResult {
@@ -168,11 +142,11 @@ export function transpile(
     };
   }
 
-  // Run function call analysis (ADR-030: define-before-use)
+  // Run call analysis (ADR-030: define-before-use)
   const funcAnalyzer = new FunctionCallAnalyzer();
   const funcErrors = funcAnalyzer.analyze(tree);
 
-  // Convert function call errors to transpile errors
+  // Convert call errors to transpile errors
   for (const funcError of funcErrors) {
     errors.push({
       line: funcError.line,
@@ -182,7 +156,7 @@ export function transpile(
     });
   }
 
-  // If there are function call errors, fail compilation
+  // If there are call errors, fail compilation
   if (errors.length > 0) {
     return {
       success: false,
@@ -293,147 +267,6 @@ export function transpile(
       declarationCount,
     };
   }
-}
-
-/**
- * Parse C-Next source and return parse result without generating code
- * Convenience wrapper around transpile with parseOnly: true
- */
-export function parse(source: string): ITranspileResult {
-  return transpile(source, { parseOnly: true });
-}
-
-/**
- * Map ESymbolKind to TSymbolKind for extension use
- */
-function mapSymbolKind(kind: ESymbolKind): TSymbolKind {
-  switch (kind) {
-    case ESymbolKind.Namespace:
-      return "namespace";
-    case ESymbolKind.Class:
-      return "class";
-    case ESymbolKind.Struct:
-      return "struct";
-    case ESymbolKind.Register:
-      return "register";
-    case ESymbolKind.Function:
-      return "function";
-    case ESymbolKind.Variable:
-      return "variable";
-    case ESymbolKind.RegisterMember:
-      return "registerMember";
-    default:
-      return "variable";
-  }
-}
-
-/**
- * Extract local name from full qualified name
- * e.g., "LED_toggle" with parent "LED" -> "toggle"
- */
-function extractLocalName(fullName: string, parent?: string): string {
-  if (parent && fullName.startsWith(parent + "_")) {
-    return fullName.substring(parent.length + 1);
-  }
-  return fullName;
-}
-
-/**
- * Parse C-Next source and extract symbols for IDE features
- *
- * Unlike transpile(), this function attempts to extract symbols even when
- * there are parse errors, making it suitable for autocomplete during typing.
- *
- * @param source - C-Next source code string
- * @returns Parse result with symbols
- *
- * @example
- * ```typescript
- * import { parseWithSymbols } from './lib/transpiler';
- *
- * const result = parseWithSymbols(source);
- * // Find namespace members for autocomplete
- * const ledMembers = result.symbols.filter(s => s.parent === 'LED');
- * ```
- */
-export function parseWithSymbols(source: string): IParseWithSymbolsResult {
-  const errors: ITranspileError[] = [];
-
-  // Create the lexer and parser
-  const charStream = CharStream.fromString(source);
-  const lexer = new CNextLexer(charStream);
-  const tokenStream = new CommonTokenStream(lexer);
-  const parser = new CNextParser(tokenStream);
-
-  // Custom error listener to collect errors
-  lexer.removeErrorListeners();
-  parser.removeErrorListeners();
-
-  const errorListener = {
-    syntaxError(
-      _recognizer: unknown,
-      _offendingSymbol: unknown,
-      line: number,
-      charPositionInLine: number,
-      msg: string,
-      _e: unknown,
-    ): void {
-      errors.push({
-        line,
-        column: charPositionInLine,
-        message: msg,
-        severity: "error",
-      });
-    },
-    reportAmbiguity(): void {},
-    reportAttemptingFullContext(): void {},
-    reportContextSensitivity(): void {},
-  };
-
-  lexer.addErrorListener(errorListener);
-  parser.addErrorListener(errorListener);
-
-  // Parse the input - continue even with errors to get partial symbols
-  let tree;
-  try {
-    tree = parser.program();
-  } catch (e) {
-    const errorMessage = e instanceof Error ? e.message : String(e);
-    errors.push({
-      line: 1,
-      column: 0,
-      message: `Parse failed: ${errorMessage}`,
-      severity: "error",
-    });
-    return {
-      success: false,
-      errors,
-      symbols: [],
-    };
-  }
-
-  // Collect symbols from the parse tree
-  const collector = new CNextSymbolCollector("<source>");
-  const rawSymbols = collector.collect(tree);
-
-  // Transform ISymbol[] to ISymbolInfo[]
-  const symbols: ISymbolInfo[] = rawSymbols.map((sym) => ({
-    name: extractLocalName(sym.name, sym.parent),
-    fullName: sym.name,
-    kind: mapSymbolKind(sym.kind),
-    type: sym.type,
-    parent: sym.parent,
-    signature: sym.signature,
-    accessModifier: sym.accessModifier,
-    line: sym.sourceLine,
-    size: sym.size,
-  }));
-
-  return {
-    success: errors.length === 0,
-    errors,
-    symbols,
-  };
 }
 
 export default transpile;
