+++ src/lib/types/IParseWithSymbolsResult.ts
@@ -1,0 +1,16 @@
+import ITranspileError from "./ITranspileError";
+import ISymbolInfo from "./ISymbolInfo";
+
+/**
+ * Result of parsing with symbol extraction
+ */
+interface IParseWithSymbolsResult {
+  /** Whether parsing succeeded without errors */
+  success: boolean;
+  /** List of errors and warnings */
+  errors: ITranspileError[];
+  /** Extracted symbols */
+  symbols: ISymbolInfo[];
+}
+
+export default IParseWithSymbolsResult;
