+++ src/lib/types/TSymbolKind.ts
@@ -1,0 +1,15 @@
+/**
+ * Symbol kind for IDE features
+ */
+type TSymbolKind =
+  | "namespace"
+  | "class"
+  | "struct"
+  | "register"
+  | "function"
+  | "variable"
+  | "registerMember"
+  | "field"
+  | "method";
+
+export default TSymbolKind;
