+++ src/preprocessor/types/IPreprocessOptions.ts
@@ -1,0 +1,20 @@
+import IToolchain from "./IToolchain";
+
+/**
+ * Preprocessor options
+ */
+interface IPreprocessOptions {
+  /** Additional include paths */
+  includePaths?: string[];
+
+  /** Preprocessor defines (-D flags) */
+  defines?: Record<string, string | boolean>;
+
+  /** Specific toolchain to use (auto-detect if not specified) */
+  toolchain?: IToolchain;
+
+  /** Keep #line directives for source mapping (default: true) */
+  keepLineDirectives?: boolean;
+}
+
+export default IPreprocessOptions;
