+++ src/project/types/IProjectConfig.ts
@@ -1,8 +1,4 @@
 /**
- * Project Configuration Types
- */
-
-/**
  * Configuration for a C-Next project
  */
 interface IProjectConfig {
@@ -34,31 +30,4 @@ interface IProjectConfig {
   defines?: Record<string, string | boolean>;
 }
 
-/**
- * Result of compiling a project
- */
-interface IProjectResult {
-  /** Whether compilation succeeded */
-  success: boolean;
-
-  /** Number of files processed */
-  filesProcessed: number;
-
-  /** Number of symbols collected */
-  symbolsCollected: number;
-
-  /** Symbol conflicts detected */
-  conflicts: string[];
-
-  /** Errors encountered */
-  errors: string[];
-
-  /** Warnings */
-  warnings: string[];
-
-  /** Generated output files */
-  outputFiles: string[];
-}
-
 export default IProjectConfig;
-export type { IProjectResult };
